rpc: fix deadlock caused by priority inversion
The priority inversion occured when userspace process attempting to
send a message and was pre-empted after message reservation by the
workqueue processing prior finishing the send.

Fix by changing the rpc communication to use the new MPMC algorithm and
replacing the problematic message state with an additional queue that
the kernel module uses to forward the non-ioeventfd messages for the
userspace.

Signed-off-by: Markku Ahvenjärvi <mankku@gmail.com>